Move sample to poly.rs
