+++ eyelid-match-ops/src/primitives/poly.rs
@@ -5,6 +5,7 @@
 
 use ark_ff::{Field, One, Zero};
 use ark_poly::polynomial::Polynomial;
+use rand::Rng;
 
 pub use fq::Coeff;
 pub use modular_poly::{
@@ -120,4 +121,19 @@ pub fn extended_gcd<const MAX_POLY_DEGREE: usize>(
     }
 
     (x_prev, y_prev, ri_prev)
+}
+
+/// This sampling is similar to what will be necessary for YASHE KeyGen
+/// TODO: generate Gaussian distribution instead of "uniform"
+pub fn sample<const MAX_POLY_DEGREE: usize>() -> Poly<MAX_POLY_DEGREE> {
+    let mut rng = rand::thread_rng();
+    let mut res = Poly::zero();
+    let max_coeff = 8;
+    let t = 2;
+    for i in 0..MAX_POLY_DEGREE {
+        let coeff_rand = rng.gen_range(1..max_coeff);
+        res[i] = Coeff::from(t * coeff_rand);
+    }
+    res[0] += Coeff::one();
+    res
 }
